Unified-AI/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
torch
transformers
sentence-transformers
openai-whisper
numpy
Pillow
groq
requests
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
langgraph
tavily-python>=0.3.0
httpx
pydantic
